requirements.txt
faiss_cpu==1.11.0.post1
numpy==2.3.2
pdfplumber==0.11.7
sentence_transformers==5.0.0
tqdm==4.67.1
llama-cpp-python